+++ R/reexport.R
@@ -1,5 +1,0 @@
-#' @importFrom rxode2 %>%
-#' @export
-rxode2::`%>%`
-
-